+++ include/engine/wavefront/WavefrontRaytracer.h
@@ -73,6 +73,7 @@ namespace engine::wavefront {
       bool intersectionBackendPlatformGpuRenderPathAvailable = false;
       std::uint64_t intersectionBackendExpectedRays = 0;
       std::uint64_t intersectionBackendAutoMinimumGpuRays = 0;
+      std::uint64_t intersectionBackendAutoEstimatedQueryTransferBytes = 0;
       bool intersectionSceneCompiled = false;
       std::uint64_t intersectionSceneBvhNodes = 0;
       std::uint64_t intersectionScenePrimitives = 0;